Time filters and OS uptime on the high-resolution counter
Time filters (One Euro, Damping, Speed) measured their interval on the
32-bit millisecond counter, so at its wrap, every 49.7 days of uptime,
one step read an interval of 0. One Euro divided by it (1 / 0, then 0 * inf
= NaN for an unchanged value) and kept the NaN in its state, so its output
stayed stuck from then on. They use the high-resolution counter now, and
One Euro skips a zero step. Their start times were also set before the
array was sized, so they stayed 0 and a filter's first interval was the
machine's uptime.

The OS module's System and Process Up Time read the same counter (the
process time went negative after 49.7 days): high-resolution counter, in
doubles.

+++ Source/Module/modules/system/os/OSModule.h
@@ -49,7 +49,7 @@ public:
 	ControllableContainer pingIPsCC;
 	ControllableContainer pingStatusCC;
 
-	static float timeAtProcessStart;
+	static double timeAtProcessStart;
 
 	//Script
 	const Identifier launchAppId = "launchApp";
